KIEKER-1790 Reworked multi connection tcp stage to cost less CPU time.
