+++ kieker-tools/collector/collector.properties
@@ -23,7 +23,7 @@ kieker.monitoring.metadata=true
 ## define input
 # TCP server for multiple connections
 kieker.tools.source=kieker.tools.source.MultipleConnectionTcpSourceCompositeStage
-kieker.tools.source.MultipleConnectionTcpSourceCompositeStage.port=9876
+kieker.tools.source.MultipleConnectionTcpSourceCompositeStage.port=9879
 kieker.tools.source.MultipleConnectionTcpSourceCompositeStage.capacity=8192
 
 # # File input
